fix(mediatype,pipeline): deterministic Header lookup, doc + test-file alignment (LAB-4678)
Address CodeRabbit review comments on PR #188:
- Header: exact-key match first, then lexicographically-smallest matching
  key as a stable fallback, so a capture holding both content-type and
  Content-Type resolves the same value run-to-run. Add collision regression test.
- Extract logClassificationReasons into reasons.go so reasons_test.go matches
  its source file; prune now-unused pipeline.go imports.
- Document the exported Header API in mediatype doc.go.
